+++ api/Cargo.toml
@@ -4,3 +4,5 @@ version = "0.1.0"
 edition = "2021"
 
 [dependencies]
+registry.workspace = true
+axum.workspace = true
